// types/CommonInputTextProps.ts
export interface CommonInputTextProps extends InputHTMLAttributes<HTMLInputElement> {
    label?: string;
    labelText?: string;
    labelClassName?: string;
    value?: string | number | undefined;
    name?: string;
    className?: any;
    type?: string;
    autoFocus?: boolean;
    borderRadius?: string;
    placeholder?: string;
    autoComplete?: string;
    required?: boolean;
    disabled?: boolean;
    readOnly?: boolean;
    error?: any;
    helperText?: string | false | string[] | FormikErrors<any> | FormikErrors<any>[] | undefined;
    multiline?: boolean;
    rows?: number;
    maxRows?: number;
    color?: any;
    showIcon?: any;
    borderColor?: string;
    focusedBorderColor?: string;
    labelColor?: string;
    focusColor?: string;
    bgColor?: string;
    inputWidth?: string,
    textFieldSize?: any;
    endAdornment?: any;
    startAdornment?: any;
    onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
    onKeyPress?: (event: React.KeyboardEvent<HTMLInputElement>) => void;
}